fix: renamed view
