+++ apps/cranio-provider/src/views/provider-cs-center-general.vue
